fix!: fixed refactor except StateManager which is still bork
