+++ backend/src/main/java/se/cholmers/backend/BackendApplication.java
@@ -3,6 +3,8 @@ package se.cholmers.backend;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+import se.cholmers.backend.Model.StateManager;
+
 @SpringBootApplication
 public class BackendApplication {
