+++ backend/src/main/java/se/cholmers/backend/Model/Cart.java
@@ -1,6 +1,7 @@
 package se.cholmers.backend.Model;
 
 import java.util.HashMap;
+import java.util.Map;
 
 public class Cart {
     private Map<Product, Integer> itemsInCart;
