+++ backend/src/main/java/se/cholmers/backend/Model/Order.java
@@ -1,6 +1,7 @@
 package se.cholmers.backend.Model;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 class Order {
     private LocalDateTime timeStamp;
